server/requirement.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.3.1    
anyio==4.2.0
asttokens==2.4.1
astunparse==1.6.3        
async-timeout==4.0.3     
attrs==23.2.0
beautifulsoup4==4.12.2   
blinker==1.4
blis==0.7.11
catalogue==2.0.10        
certifi==2023.11.17      
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.36.0       
cloudpathlib==0.16.0     
cog==0.8.6
coloredlogs==15.0.1      
command-not-found==0.3   
confection==0.1.4        
contourpy==1.2.0
cryptography==3.4.8      
cycler==0.12.1
cymem==2.0.8
dbus-python==1.2.18      
decorator==5.1.1
dill==0.3.7
discord.py==2.3.2
distro==1.7.0
distro-info===1.1build1
exceptiongroup==1.2.0
executing==2.0.1
fastai==2.7.14
fastapi==0.98.0
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.13.1
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.5.4
gdown==4.7.1
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httplib2==0.20.2
httptools==0.6.1
humanfriendly==10.0
idna==3.6
imageio==2.33.1
imgviz==1.7.4
importlib-metadata==4.6.4
ipython==8.19.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
keras==3.1.1
keyring==23.5.0
kiwisolver==1.4.5
labelme==5.3.1
langcodes==3.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
libclang==18.1.1
lxml==4.9.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
namex==0.0.7
natsort==8.4.0
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
onnxruntime==1.16.3
opencv-python==4.7.0.72
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pdf2image==1.16.3
pexpect==4.9.0
Pillow==10.1.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pybboxes==0.1.6
pydantic==1.10.13
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==5.4.1
QtPy==2.4.1
qudida==0.0.4
requests==2.31.0
rich==13.7.1
sahi==0.11.15
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
SecretStorage==3.3.1
selenium==4.17.2
shapely==2.0.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
structlog==23.3.0
sympy==1.12
systemd-python==234
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminaltables==3.1.10
thinc==8.2.3
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.12.9
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.14.0
trio==0.24.0
trio-websocket==0.11.1
triton==2.1.0
typer==0.9.4
typing_extensions==4.9.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
ultralytics==8.0.229
unattended-upgrades==0.1
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
wadllib==1.3.6
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.12
weasel==0.3.4
webdriver-manager==4.0.1
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==1.0.0